alterado parametro no repository para long, adicionado endpoint de attexperiencia no controller, e criado metodo de attexperiencia no service

+++ src/main/java/br/unipar/devbackend/fixr/Repository/EstatisticasPrestadorRepository.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 @Repository
 public interface EstatisticasPrestadorRepository extends JpaRepository<EstatisticasPrestador, Long> {
 
-    Optional<EstatisticasPrestador> findByPrestadorId(Integer prestadorId);
+    Optional<EstatisticasPrestador> findByPrestadorId(Long prestadorId);
 
 }
